+++ src/main/java/com/nhnacademy/store99/bookstore/order_book/DTO/response/BestBookResponse.java
@@ -1,0 +1,4 @@
+package com.nhnacademy.store99.bookstore.order_book.DTO.response;
+
+public class BestBookResponse {
+}
