+++ src/main/java/com/nhnacademy/store99/bookstore/order_book/repository/OrderBookRepository.java
@@ -1,6 +1,6 @@
 package com.nhnacademy.store99.bookstore.order_book.repository;
 
-import com.nhnacademy.store99.bookstore.order_book.response.LatestBookResponse;
+import com.nhnacademy.store99.bookstore.order_book.DTO.response.LatestBookResponse;
 import java.util.List;
 import java.util.Map;
