--- src/main/java/com/nhnacademy/store99/bookstore/order_book/response/BestBookResponse.java
@@ -1,4 +1,0 @@
-package com.nhnacademy.store99.bookstore.order_book.response;
-
-public class BestBookResponse {
-}
