--- src/main/java/com/nhnacademy/store99/bookstore/order_book/response/LatestBookResponse.java
@@ -1,4 +1,0 @@
-package com.nhnacademy.store99.bookstore.order_book.response;
-
-public class LatestBookResponse {
-}
