+++ src/main/java/com/nhnacademy/store99/bookstore/order_book/service/OrderBookService.java
@@ -1,6 +1,6 @@
 package com.nhnacademy.store99.bookstore.order_book.service;
 
-import com.nhnacademy.store99.bookstore.order_book.response.LatestBookResponse;
+import com.nhnacademy.store99.bookstore.order_book.DTO.response.LatestBookResponse;
 import java.util.List;
 import java.util.Map;
