+++ src/main/java/com/nhnacademy/store99/bookstore/order_book/service/impl/OrderBookServiceImpl.java
@@ -1,7 +1,7 @@
 package com.nhnacademy.store99.bookstore.order_book.service.impl;
 
+import com.nhnacademy.store99.bookstore.order_book.DTO.response.LatestBookResponse;
 import com.nhnacademy.store99.bookstore.order_book.repository.OrderBookRepository;
-import com.nhnacademy.store99.bookstore.order_book.response.LatestBookResponse;
 import com.nhnacademy.store99.bookstore.order_book.service.OrderBookService;
 import java.util.List;
 import java.util.Map;
